♻️ UserError and userAssert without runtime + shared parseDateAttrs helper (#32407)
* Define UserError for use in Preact components

* Add parseDateAttrs helper to src/utils/date

* jsdoc lint fix

* add userAssert helper for use in preact

* Use shared parseDateAttrs in date-display, date-countdown, and timeago

* un-export parseEpoch

* Drop unused import

* Remove unsupported jsdoc comment

* clean up super-duped parseDateAttrs tests

* Remove unused imports

* @closurePrimitive


* no Number as callback fn


* make pure(User|Dev)Assert and validate attributes

* String-only assertions

* fix type annotations for assertions

* Import amp components in tests

* Import sentinel from pure-assert for now

* Re-export

* remove closurePrimitive on pureAssertion
