+++ extensions/amp-timeago/1.0/test/test-amp-timeago.js
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import {parseDateAttrs} from '../amp-timeago';
+import '../amp-timeago';
 import {toggleExperiment} from '../../../../src/experiments';
 import {waitFor} from '../../../../testing/test-helper.js';
 
@@ -116,70 +116,3 @@ describes.realWin(
     });
   }
 );
-
-describes.sandboxed('amp-date-display 1.0: parseDateAttrs', {}, (env) => {
-  const DATE = new Date(1514793600000);
-  const DATE_STRING = DATE.toISOString();
-
-  let element;
-
-  beforeEach(() => {
-    element = document.createElement('amp-date-display');
-  });
-
-  it('should throw when no date is specified', () => {
-    expect(() => parseDateAttrs(element)).to.throw(/required/);
-  });
-
-  it('should throw when invalid date is specified', () => {
-    element.setAttribute('datetime', 'invalid');
-    expect(() => parseDateAttrs(element)).to.throw(/Invalid date/);
-  });
-
-  it('should parse the "datetime" attribute', () => {
-    element.setAttribute('datetime', DATE_STRING);
-    expect(parseDateAttrs(element)).to.equal(DATE.getTime());
-
-    // With offset.
-    element.setAttribute('offset-seconds', '1');
-    expect(parseDateAttrs(element)).to.equal(DATE.getTime() + 1000);
-  });
-
-  it('should accept "datetime=now"', () => {
-    env.sandbox.useFakeTimers(DATE);
-    element.setAttribute('datetime', 'now');
-    expect(parseDateAttrs(element)).to.equal(DATE.getTime());
-
-    // With offset.
-    element.setAttribute('offset-seconds', '1');
-    expect(parseDateAttrs(element)).to.equal(DATE.getTime() + 1000);
-  });
-
-  it('should parse the "timestamp-ms" attribute', () => {
-    element.setAttribute('timestamp-ms', DATE.getTime());
-    expect(parseDateAttrs(element)).to.equal(DATE.getTime());
-
-    // With offset.
-    element.setAttribute('offset-seconds', '1');
-    expect(parseDateAttrs(element)).to.equal(DATE.getTime() + 1000);
-  });
-
-  it('should throw when invalid "timestamp-ms" is specified', () => {
-    element.setAttribute('timestamp-ms', 'invalid');
-    expect(() => parseDateAttrs(element)).to.throw(/required/);
-  });
-
-  it('should parse the "timestamp-seconds" attribute', () => {
-    element.setAttribute('timestamp-seconds', DATE.getTime() / 1000);
-    expect(parseDateAttrs(element)).to.equal(DATE.getTime());
-
-    // With offset.
-    element.setAttribute('offset-seconds', '1');
-    expect(parseDateAttrs(element)).to.equal(DATE.getTime() + 1000);
-  });
-
-  it('should throw when invalid "timestamp-seconds" is specified', () => {
-    element.setAttribute('timestamp-seconds', 'invalid');
-    expect(() => parseDateAttrs(element)).to.throw(/required/);
-  });
-});
